LayoutTests/pdf/pdf-plugin-printing-expected.txt

Basic testing for automatic PDF printing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Received beforeprint event
* Received afterprint event
PASS receivedBeforePrintEvent is true
PASS successfullyParsed is true

TEST COMPLETE
